Use official mongodb driver
I used the mgo driver (because it's easier to use) but it doesn't
support the newer version of mongodb (4+) and also doesn't support the
newer connection string format used exclusively by mongodb atlas
(mongo's cloud offering for mongodb, its freee soooooooooo, we are totes
gonna support it).

+++ pkg/storage/migrations/manager.go
@@ -118,7 +118,7 @@ func (m *Manager) Aggregate(collection string, opts storage.AggregateOptions, ou
 	return m.store.Aggregate(collection, opts, out)
 }
 
-func (m *Manager) Count(collection string, opts storage.CountOptions) (int, error) {
+func (m *Manager) Count(collection string, opts storage.CountOptions) (int64, error) {
 	if err := m.Connect(); err != nil {
 		return 0, err
 	}
